Add files for reducer, middleware and initialState
